services/apis/django-api/source/requirements.txt
# Django itself plus addon packages. 
Django==3.1.*
channels==3.0.*
django-guardian==2.3.*
djangorestframework==3.12.*
drf-spectacular==0.11.*

# Required for communication with other services.
paho-mqtt==1.5.*

# To allow running the tests.
pytest-django

# To parse settings from .env files.
python-dotenv==0.15.*

# Production servers.
daphne==3.0.*
whitenoise==5.2.*
